- Fixed Paymium orderbook implementation - Disabled BxInTh as it is a localised provider, not international - Implemented GetOrderBookAsync in AbucoinsProvider - Implemented TestGetOrderBook in AbucoinsTests - Implemented GetOrderBookAsync in BitfinexProvider - Implemented TestGetOrderBook in BitfinexTests - Implemented GetOrderBookAsync in CoinroomProvider - Implemented TestGetOrderBook in CoinroomTests - Implemented GetOrderBookAsync in LiquiProvider - Implemented TestGetOrderBook in LiquiTests - Implemented GetOrderBookAsync in TuxExchangeProvider - Implemented TestGetOrderBook in TuxExchangeTests - Implemented GetOrderBookAsync in GateProvider - Implemented TestGetOrderBook in GateTests - Implemented GetOrderBookAsync in BleutradeProvider - Implemented TestGetOrderBook in BleutradeTests

+++ Prime.Plugins/Services/Bitfinex/BitfinexSchema.cs
@@ -214,7 +214,7 @@ namespace Prime.Plugins.Services.Bitfinex
         {
             public decimal price;
             public decimal amount;
-            public long timestamp;
+            public decimal timestamp;
         }
 
         internal class OrderBookResponse
